Renaming logins into auth
