+++ index.js
@@ -9,7 +9,7 @@ const customerRoute = require('./routers/customers');
 const movieRoute = require('./routers/movies');
 const rentalRoute = require('./routers/rentals');
 const userRoute = require('./routers/users');
-const loginRoute = require('./routers/logins');
+const authRoute = require('./routers/auth');
 const app = express();
 const mongoose = require('mongoose');
 const Joi = require('joi');
@@ -41,6 +41,6 @@ app.use('/api/customers', customerRoute);
 app.use('/api/movies', movieRoute);
 app.use('/api/rentals', rentalRoute);
 app.use('/api/users', userRoute);
-app.use('/api/logins', loginRoute);
+app.use('/api/auth', authRoute);
 
 app.listen(port, () => console.log(`Listening on port ${port}...`));
